Refactor carousel logic and make it a hook
